remove board.cpp and board.hpp, board is so simple that can be included in aruco.hpp
